requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
yfinance==0.2.33
joblib==1.3.2
numpy==1.26.2
pydantic==2.5.2
requests==2.31.0